refactor(ses): fold VDM into SesAccountService (strangler step 7)
VDM (accountVdmStore + the account-level VdmAttributes methods) entered the
tree via the merged PutAccountVdmAttributes work and was left on the facade
during the rebase. It is account-level with no shared helper, so move it into
SesAccountService alongside the sending-enabled flag — where it was always meant to live. The facade delegates find/putAccountVdmAttributes;
SesController's getAccount VdmAttributes rendering is unchanged.
