+++ src/test/java/io/github/hectorvent/floci/services/ses/SesAccountServiceTest.java
@@ -1,6 +1,7 @@
 package io.github.hectorvent.floci.services.ses;
 
 import io.github.hectorvent.floci.core.storage.InMemoryStorage;
+import io.github.hectorvent.floci.services.ses.model.AccountVdmAttributes;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
@@ -8,8 +9,9 @@ import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 /**
- * Unit test for the extracted account domain. Like the receipt-rule step, the
- * service is constructed with just its own store — no 14-argument SesService needed.
+ * Unit test for the extracted account domain: the account sending-enabled flag and
+ * the VDM attributes. The service is constructed with just its own two stores — no 14-argument
+ * SesService needed.
  */
 class SesAccountServiceTest {
 
@@ -18,7 +20,7 @@ class SesAccountServiceTest {
 
     @BeforeEach
     void setUp() {
-        service = new SesAccountService(new InMemoryStorage<>());
+        service = new SesAccountService(new InMemoryStorage<>(), new InMemoryStorage<>());
     }
 
     @Test
@@ -39,5 +41,17 @@ class SesAccountServiceTest {
     void sendingEnabled_isPerRegion() {
         service.setAccountSendingEnabled(REGION, false);
         assertTrue(service.isAccountSendingEnabled("eu-west-1"));
+    }
+
+    @Test
+    void vdmAttributes_absentUntilConfigured_thenRoundTrip() {
+        // Opt-in: a never-configured region has no VdmAttributes at all.
+        assertTrue(service.findAccountVdmAttributes(REGION).isEmpty());
+
+        service.putAccountVdmAttributes(REGION, new AccountVdmAttributes(true, true, false));
+        AccountVdmAttributes vdm = service.findAccountVdmAttributes(REGION).orElseThrow();
+        assertTrue(vdm.vdmEnabled());
+        assertTrue(vdm.engagementMetrics());
+        assertFalse(vdm.optimizedSharedDelivery());
     }
 }
