+++ src/test/java/io/github/hectorvent/floci/services/ses/SesServiceSuppressionLegacyKeyTest.java
@@ -48,11 +48,10 @@ class SesServiceSuppressionLegacyKeyTest {
         service = new SesService(
                 new InMemoryStorage<String, Identity>(),
                 new InMemoryStorage<String, SentEmail>(),
-                new SesAccountService(new InMemoryStorage<String, Boolean>()),
+                new SesAccountService(new InMemoryStorage<String, Boolean>(), new InMemoryStorage<String, AccountVdmAttributes>()),
                 new InMemoryStorage<String, EmailTemplate>(),
                 new InMemoryStorage<String, ConfigurationSet>(),
                 new SesSuppressionService(suppressionStore, new InMemoryStorage<String, AccountSuppressionAttributes>()),
-                new InMemoryStorage<String, AccountVdmAttributes>(),
                 new InMemoryStorage<String, DedicatedIpPool>(),
                 new SesContactService(new InMemoryStorage<String, ContactList>(), new InMemoryStorage<String, Contact>(), Clock.systemUTC()),
                 new SesPolicyService(new InMemoryStorage<String, String>(), new ObjectMapper()),
